Update typechecking for tests and fix some type errors. Finally remove all references to Foreign

+++ tests/LazyOperators.hs
@@ -3,5 +3,4 @@ import           Prelude
 main :: Fay ()
 main = print testFn
 
-testFn = let f a b = snd (a/b,10) in f 1 0 -- undefined undefined
-
+testFn = let f a b = snd (a/b,10::Double) in f 1 0 -- undefined undefined
